refactor(gda): lift the import-evidence adapter into a core import_evidence module
The faithful mirror of `EditorFileSystem::_test_for_reimport` that
`resource import` uses to decide `cached` / `missing` / `stale` / `invalid`
lived as module-level regexes and private helpers inside the resource group
module, so its only interface was the CLI. Move it to `gda/import_evidence.py`:
the group keeps its whole command slice (ADR-0040) and the core hosts the
non-command engine knowledge, as `binary` and `display` already do.

The module owns the receipt grammar, the sidecar/receipt regexes,
`asset_state`, `project_import_gaps`, `CACHE_ROOT_REL`, and
`classify_created_file` — the one-line "under the cache root" rule the
command's created-file list applied inline. Its docstring carries the
parity contract moved intact, remainder and direction included.

`asset_state` returns the narrowed `AssetEvidence`, whose `EvidenceStatus` is
four-valued: a settlement stays the command's post-pass mapping onto the wire
`ResourceImportAsset`. Request normalization and created-file accounting stay
in the command. Dependency direction is commands -> core only.

Behavior-preserving: `gda schema` and every `--help` page byte-identical to
the base (sha256 abb4a866… and d3c64bf5…).

Refs #741
